pyproject.toml
[project]
name = "pwlsplit"
version = "1.0.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Will Zhang", email = "willwz@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
    "matplotlib",
    "scipy-stubs[scipy]",
    "pytools@git+https://github.com/willwiz/pytools",
    "pandas-stubs",
    "pandas",
]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
pytools = [{ path = ".deps/pytools", editable = true }]
